cli_copilot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.1
pydantic==2.5.2
typing-extensions==4.8.0
python-dotenv==1.0.0